+++ nugurang-front/components/BottomNavBar.js
@@ -3,12 +3,13 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import BottomNavigation from '@material-ui/core/BottomNavigation';
 import BottomNavigationAction from '@material-ui/core/BottomNavigationAction';
-import HomeIcon from '@material-ui/icons/Home';
+import DashboardIcon from '@material-ui/icons/Dashboard';
 import ForumIcon from '@material-ui/icons/Forum';
 import ChatIcon from '@material-ui/icons/ChatBubble';
 import NotificationsIcon from '@material-ui/icons/Notifications';
+import BusinessCenterIcon from '@material-ui/icons/BusinessCenter';
 import PersonIcon from '@material-ui/icons/Person';
-import PeopleIcon from '@material-ui/icons/People';
+import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
 import AppBar from '@material-ui/core/AppBar';
 
 const useStyles = makeStyles((theme) => ({
@@ -29,11 +30,11 @@ export default function BottomNavBar() {
   return (
     <AppBar className={classes.appBar}>
       <BottomNavigation value={value} onChange={handleChange} className={classes.root}>
-        <BottomNavigationAction label="Home" value="home" icon={<HomeIcon />} href="/"/>
+        <BottomNavigationAction label="Dashboard" value="dashboard" icon={<DashboardIcon />} href="/"/>
         <BottomNavigationAction label="Forum" value="forum" icon={<ForumIcon />} />
-        <BottomNavigationAction label="People" value="people" icon={<PeopleIcon />} />
+        <BottomNavigationAction label="Team" value="team" icon={<BusinessCenterIcon />} />
         <BottomNavigationAction label="Person" value="person" icon={<PersonIcon />} />
-        <BottomNavigationAction label="Sign in" value="signin" icon={<PersonIcon />} href="/signin"/>
+        <BottomNavigationAction label="Sign in" value="signin" icon={<MoreHorizIcon />} href="/signin"/>
       </BottomNavigation>
     </AppBar>
   );
