--- nugurang-front/components/Layout.js
@@ -1,16 +1,0 @@
-import BottomNavBar from "./BottomNavBar";
-
-const layoutStyle = {
-  margin: 20,
-  padding: 20,
-  border: "1px solid #DDD"
-};
-
-const Layout = props => (
-  <div style={layoutStyle}>
-    {props.children}
-    <BottomNavBar />
-  </div>
-);
-
-export default Layout;
