+++ nugurang-front/pages/signin.js
@@ -13,12 +13,12 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import AppBar from '@material-ui/core/AppBar';
-import Layout from '../components/Layout'
 
 
 const useStyles = makeStyles((theme) => ({
   paper: {
     marginTop: theme.spacing(8),
+    marginBottom: theme.spacing(12),
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center',
@@ -39,67 +39,67 @@ const useStyles = makeStyles((theme) => ({
 export default function SignIn() {
   const classes = useStyles();
   return (
-    <Layout>
-      <Container component="main" maxWidth="xs">
-        <CssBaseline />
-        <div className={classes.paper}>
-          <Avatar className={classes.avatar}>
-            <LockOutlinedIcon />
-          </Avatar>
-          <Typography component="h1" variant="h5">
-            Sign in
-          </Typography>
-          <form className={classes.form} noValidate>
-            <TextField
-              variant="outlined"
-              margin="normal"
-              required
-              fullWidth
-              id="email"
-              label="Email Address"
-              name="email"
-              autoComplete="email"
-              autoFocus
-            />
-            <TextField
-              variant="outlined"
-              margin="normal"
-              required
-              fullWidth
-              name="password"
-              label="Password"
-              type="password"
-              id="password"
-              autoComplete="current-password"
-            />
-            <FormControlLabel
-              control={<Checkbox value="remember" color="primary" />}
-              label="Remember me"
-            />
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-            >
-              Sign In
-            </Button>
-            <Grid container>
-              <Grid item xs>
-                <Link href="#" variant="body2">
-                  Forgot password?
-                </Link>
-              </Grid>
-              <Grid item>
-                <Link href="#" variant="body2">
-                  {"Don't have an account? Sign Up"}
-                </Link>
-              </Grid>
-            </Grid>
-          </form>
-        </div>
-      </Container>
-    </Layout>
+
+	  <Container component="main" maxWidth="xs">
+	    <CssBaseline />
+	    <div className={classes.paper}>
+	      <Avatar className={classes.avatar}>
+	        <LockOutlinedIcon />
+	      </Avatar>
+	      <Typography component="h1" variant="h5">
+	        Sign in
+	      </Typography>
+	      <form className={classes.form} noValidate>
+	        <TextField
+	          variant="outlined"
+	          margin="normal"
+	          required
+	          fullWidth
+	          id="email"
+	          label="Email Address"
+	          name="email"
+	          autoComplete="email"
+	          autoFocus
+	        />
+	        <TextField
+	          variant="outlined"
+	          margin="normal"
+	          required
+	          fullWidth
+	          name="password"
+	          label="Password"
+	          type="password"
+	          id="password"
+	          autoComplete="current-password"
+	        />
+	        <FormControlLabel
+	          control={<Checkbox value="remember" color="primary" />}
+	          label="Remember me"
+	        />
+	        <Button
+	          type="submit"
+	          fullWidth
+	          variant="contained"
+	          color="primary"
+	          className={classes.submit}
+	        >
+	          Sign In
+	        </Button>
+	        <Grid container>
+	          <Grid item xs>
+	            <Link href="#" variant="body2">
+	              Forgot password?
+	            </Link>
+	          </Grid>
+	          <Grid item>
+	            <Link href="/signup" variant="body2">
+	              {"Don't have an account? Sign Up"}
+	            </Link>
+	          </Grid>
+	        </Grid>
+	      </form>
+	    </div>
+	  </Container>
+
   );
 }
